run zokrates asynchronously
Because ZoKrates writes multiple files to disk instead of returning all output
to stdout,
only one ZoKrates can run at any time.
To work around this we use the actor pattern: computation requests are enqueued,
if ZoKrates is idle we start it right away,
and every time ZoKrates is done we check if there is more work to be started.